requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.2.4
python-dotenv==1.1.0
